requirements.txt
cassandra-driver>=3.29.0
python-dotenv==1.0.0
pandas>=2.0.0
jupyter>=1.0.0
ipython>=8.0.0
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
flake8==6.1.0
black==23.12.0
